requirements.txt
openai-whisper
sounddevice
numpy
pynput
pyperclip
Pillow
